+++ dungeon-game-core/src/com/dungeon/game/entity/Drop.java
@@ -44,7 +44,7 @@ public class Drop extends Static {
 					slot.item = world.player.inv.addItem(slot.item);
 				}
 				else if(world.mouse.rb_pressed && slot.item != null) {
-					Window temp = new DescWindow(world, (int)world.mouse.x, (int)world.mouse.y, slot.item);
+					Window temp = new DescWindow(world, world.mouse.x, world.mouse.y, slot.item);
 					temp.open();
 				}
 			}
